requirements.txt
cbpro
avro-python3
aiokafka
faust
python-rocksdb